requirements.txt
numpy
pandas
scikit-learn
python-dotenv
mysql-connector-python
pymysql
mysql
# -e .